updated orientation detection
