+++ packages/legacy/app/android/app/src/main/java/com/ariesbifold/MainActivity.java
@@ -8,6 +8,18 @@ import com.facebook.react.defaults.DefaultReactActivityDelegate;
 public class MainActivity extends ReactActivity {
 
   /**
+   * Detects changes in device orientation and sends them to JavaScript by broadcasting an event.
+   * This is used to support the camera on different tablet orientations.
+   */
+  @Override
+  public void onConfigurationChanged(Configuration newConfig) {
+      super.onConfigurationChanged(newConfig);
+      Intent intent = new Intent("onConfigurationChanged");
+      intent.putExtra("newConfig", newConfig);
+      this.sendBroadcast(intent);
+  }
+
+  /**
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
