+++ packages/legacy/app/package.json
@@ -68,6 +68,7 @@
     "react-native-gifted-chat": "*",
     "react-native-keychain": "^8.1.1",
     "react-native-localize": "^2.2.4",
+    "react-native-orientation-locker": "^1.6.0",
     "react-native-permissions": "^3.8.4",
     "react-native-qrcode-svg": "^6.2.0",
     "react-native-reanimated": "^3.4.2",
